requirements.txt
openai==1.36.1
SpeechRecognition==3.10.4
pyaudio==0.2.14
gTTS==2.5.2
praat-parselmouth==0.4.3
playsound==1.2.2
python-dotenv==1.0.1
inquirer==3.3.0